INFRA-375 : Configure CircleCI to validate & deploy orbs (#1)
On the repository CircleCI-Public/circleci-orbs that contains orbs
authored by CircleCI, a different approach has been used :
https://github.com/CircleCI-Public/circleci-orbs/blob/da2e01739c2e808c6cabe2af26710ff6777b76ea/.circleci/config.yml
On each commit, it compares which files a commit is modifying and uses
this information to compute the list of the changed orbs.
This may be the optimum way to compute the list, however it is extremely
complicated to understand. See this non-official orb implementing the
actual hard stuff :
https://circleci.com/orbs/registry/orb/iynere/compare-url

I choosed a simpler approach : comparing the local checksum with the
latest online orb version published on the CircleCI registry (@volatile).

The drawback of this choice is that we can't handle multiple major
versions of the orb since the @volatile always points to latest and highest
online version.