support github package
docker image for registry that is not dockerhub are named like
"server/owner/project".
The orb support this if the context contains the DOCKER_SERVER variable.

The problem was that the publish step is done in `githubpkg` context (with the
DOCKER_SERVER variable) but the `docker save` is not (it's in the build step, so
it is `pypi` context)

We define a DOCKER_LOCAL_NAME that is the name of the local image and is not
dependant on DOCKER_SERVER